webapp/backend/requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
slowapi>=0.1.9
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
pydantic==2.7.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
email-validator>=0.0.6

# Sync Service - Sangfor SCP API
requests>=2.31.0
pycryptodome>=3.19.0

# AI Prediction - Prophet
pandas>=2.0.0
numpy>=1.24.0
prophet>=1.1.5
cachetools>=5.3.3
